refactor: move textFit to lib directory
