+++ src/lib/textFit.js
